Fix image deletion error

+++ lib/models/section.dart
@@ -88,7 +88,7 @@ class Section extends ChangeNotifier {
     }
 
     for (final original in originalItems) {
-      if (!items.contains(original)) {
+      if (!items.contains(original) && original.image.contains('firebase')) {
         try {
           final ref = _storage.refFromURL(original.image as String);
           await ref.delete();
@@ -108,11 +108,13 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items) {
-      try {
-        final ref = _storage.refFromURL(item.image as String);
-        await ref.delete();
-      } catch (e) {
-        debugPrint('falha ao deletar ${item.image}');
+      if (item.image.contains('firebase')) {
+        try {
+          final ref = _storage.refFromURL(item.image as String);
+          await ref.delete();
+        } catch (e) {
+          debugPrint('falha ao deletar ${item.image}');
+        }
       }
     }
   }
